Assign colors to tags/services/marks
When creating e.g. a tag, you're now presented with a color picker to choose the color for this tag. Wherever the tag is displayed it is surrounded with a chip of that color for fast recognition.

It's not possible to change the color through the UI yet.

The color is serialized into the state.json with the other metadata for the tags.