Ignore fast subsequent events in fsnotify
When writing a file multiple Write events would be triggered right after
each other, so start a timer per file whenever a Write or Create event
was observed and only react to the event when the timer runs out.

The timer is reset when another event is observed before it triggered.

This prevents us from trying to access the file while it's still busy.