Optimize adding streams to marks
If the mark had a lot of stream ids in the query (> 2000) adding more
streams to it would take a long time due to reparsing and optimizing the
`id:2,3,4,5,6,...` query multiple times.

This patch modifies the existing tag's Conditions instead of recreating
them from scratch making the addition almost instant.

Removal of a stream still calls `ConditionsSet.Clean()` to normalize
the query, but it's still faster than reparsing the query string.

Generated marks can gather a lot of streams fast.