Try to include partially truncated end of stream
Don't buffer the truncated websocket frame when there are
no more chunks left in the stream. Just add them verbatim.
This isn't perfect for http/2 streams, since there could be more
chunks from other http2 streams than the websocket one,
and the remaining_data stays cached, but due to the buffering
until a full event is received, we're already silently truncating
http/2 traffic when trying to decode it. If that's ever a concern,
we should fix that in the http2 decoder too.