pwntools: Collapse recvuntil + send into a single sendafter
When exporting a stream as a pwntools script, make the output a bit nicer by using the `send[line]after` functions instead of `recvuntil` + `send[line]`.

This shortens the script and makes it more readable. The data chosen to receive until is the last line including newlines to try to make it unique.