Highlight data filter matches in stream data
All data filters are highlighted in the stream data view by sending the data filters of the search query per chunk direction and matching them in the browser again.

The regexes are passed as a string and might fail to parse in javascript since they are for the goregexp engine which supports non-standard syntax. Skip highlighting in this case.

This could be improved by just sending the matching ranges instead of the regexes.