Fix referencing query data conditions before they are added
With "xdata:1 then xdata:2", we only add the condition after searching for reusable regexes, but we added the regex "1" to the structure with the known future position of the cc in dcc.conditions, but it's not yet added. So when searching for regex "2", the data structure contains a reference to the not yet added entry in dcc.conditions.

Co-authored-by: spq <spq@aachen.ccc.de>