Fix search for selection after highlighted data
The remaining data after a highlighted part in a chunk wasn't wrapped in a span containing the offset of the data from the beginning of the chunk. It defaulted to starting at offset 0 which caused the search to include data from the beginning.

Wrap the remaining data after a highlight too.

Fixes #272