Pass chunk time to converters
Keep it in the cache file too. They are stored as a list of varints marking the delta in microseconds until the previous chunk. A delta of 0 marks the end of the list.

New helpers are added to the Python converter library to allow to pass the time through unchanged and iterate the chunks while ignoring the splits on different time in the same direction.